+++ packages/editor/src/store/actions.js
@@ -35,7 +35,7 @@ import {
 } from './utils/notice-builder';
 import { unlock } from '../lock-unlock';
 import { setCanvasWidth } from './private-actions';
-import { getCanvasWidthByDeviceType } from '../utils/device-type';
+import { getDevicePreviewWidthByDeviceType } from '../utils/device-type';
 
 const { getEntitySnapshot } = unlock( coreDataPrivateApis );
 
@@ -817,7 +817,7 @@ export function setDeviceType( deviceType ) {
 		const viewportSettings = unlock(
 			registry.select( blockEditorStore )
 		).getSettings().__experimentalFeatures?.viewport;
-		const width = getCanvasWidthByDeviceType(
+		const width = getDevicePreviewWidthByDeviceType(
 			deviceType,
 			viewportSettings
 		);